flash: move TiCI estimate count into handler
