+++ dbms/src/Flash/FlashService.cpp
@@ -27,6 +27,7 @@
 #include <Flash/Disaggregated/WNEstablishDisaggTaskHandler.h>
 #include <Flash/Disaggregated/WNFetchPagesStreamWriter.h>
 #include <Flash/EstablishCall.h>
+#include <Flash/EstimateTiCICountHandler.h>
 #include <Flash/FlashService.h>
 #include <Flash/Management/ManualCompact.h>
 #include <Flash/Mpp/MPPHandler.h>
@@ -38,7 +39,6 @@
 #include <IO/IOThreadPools.h>
 #include <Interpreters/Context.h>
 #include <Interpreters/SharedContexts/Disagg.h>
-#include <Interpreters/TimezoneInfo.h>
 #include <Interpreters/executeQuery.h>
 #include <Poco/Message.h>
 #include <Server/IServer.h>
@@ -49,14 +49,11 @@
 #include <Storages/KVStore/Read/RegionException.h>
 #include <Storages/KVStore/TMTContext.h>
 #include <Storages/S3/S3Common.h>
-#include <Storages/Tantivy/TiCIRequestUtils.h>
 #include <common/logger_useful.h>
 #include <grpcpp/server_builder.h>
 #include <grpcpp/support/status.h>
 #include <grpcpp/support/status_code_enum.h>
 #include <kvproto/disaggregated.pb.h>
-#include <tici-search-lib/src/lib.rs.h>
-#include <tipb/executor.pb.h>
 
 #include <chrono>
 #include <ext/scope_guard.h>
@@ -196,39 +193,6 @@ void updateSettingsForAutoSpill(ContextPtr & context, const LoggerPtr & log)
         if (need_log_warning)
             LOG_WARNING(log, "auto spill is enabled, so per operator's memory threshold is disabled");
     }
-}
-
-TimezoneInfo buildEstimateTimezoneInfo(const coprocessor::TiCIEstimateCountRequest & request)
-{
-    TimezoneInfo timezone_info;
-    if (!request.time_zone_name().empty())
-        timezone_info.resetByTimezoneName(request.time_zone_name());
-    else
-        timezone_info.resetByTimezoneOffset(request.time_zone_offset());
-    return timezone_info;
-}
-
-tipb::FTSQueryInfo parseEstimateQueryInfo(const coprocessor::TiCIEstimateCountRequest & request)
-{
-    tipb::FTSQueryInfo query_info;
-    if (!query_info.ParseFromString(request.fts_query_info()))
-        throw TiFlashException("Failed to parse fts_query_info", Errors::Coprocessor::BadRequest);
-    if (query_info.match_expr_size() == 0)
-        throw TiFlashException("Empty TiCI estimate query expression", Errors::Coprocessor::BadRequest);
-    return query_info;
-}
-
-rust::Vec<::ShardWithRange> buildEstimateShardRanges(const coprocessor::TiCIEstimateCountRequest & request)
-{
-    rust::Vec<::ShardWithRange> shards;
-    for (const auto & shard_info : request.shard_infos())
-    {
-        shards.push_back({
-            .shard_id = shard_info.shard_id(),
-            .ranges = TS::getKeyRanges(shard_info.ranges()),
-        });
-    }
-    return shards;
 }
 } // namespace
 
@@ -931,63 +895,8 @@ grpc::Status FlashService::GetEstimateTiCICount(
     if (!check_result.ok())
         return check_result;
 
-    try
-    {
-        if (request->shard_infos_size() == 0)
-            return grpc::Status::OK;
-
-        const auto keyspace_id = RequestUtils::deriveKeyspaceID(request->context());
-        const auto fts_query_info = parseEstimateQueryInfo(*request);
-        const auto timezone_info = buildEstimateTimezoneInfo(*request);
-        auto [query, column_ids] = TS::tipbToTiCIExpr(fts_query_info.match_expr(), timezone_info);
-        (void)column_ids;
-
-        auto shard_ranges = buildEstimateShardRanges(*request);
-        const auto estimate_result = estimate_count(keyspace_id, shard_ranges, query);
-        response->set_est_count(estimate_result.estimated_total_count);
-        LOG_DEBUG(
-            log,
-            "GetEstimateTiCICount done, est_count={}, input_shards={}, available_shards={}, sampled_shards={}",
-            response->est_count(),
-            request->shard_infos_size(),
-            estimate_result.available_shards,
-            estimate_result.sampled_shards);
-    }
-    catch (const TiFlashException & e)
-    {
-        LOG_WARNING(
-            log,
-            "GetEstimateTiCICount failed with TiFlash exception: {}\n{}",
-            e.displayText(),
-            e.getStackTrace().toString());
-        response->set_other_error(e.standardText());
-    }
-    catch (const Exception & e)
-    {
-        LOG_WARNING(
-            log,
-            "GetEstimateTiCICount failed with DB exception: {}\n{}",
-            e.message(),
-            e.getStackTrace().toString());
-        response->set_other_error(e.message());
-    }
-    catch (const pingcap::Exception & e)
-    {
-        LOG_WARNING(log, "GetEstimateTiCICount failed with KV exception: {}", e.message());
-        response->set_other_error(e.message());
-    }
-    catch (const std::exception & e)
-    {
-        LOG_WARNING(log, "GetEstimateTiCICount failed: {}", e.what());
-        response->set_other_error(e.what());
-    }
-    catch (...)
-    {
-        LOG_WARNING(log, "GetEstimateTiCICount failed with unknown exception");
-        response->set_other_error("other exception");
-    }
-
-    return grpc::Status::OK;
+    EstimateTiCICountHandler handler(request, response, log->identifier());
+    return handler.execute();
 }
 
 grpc::Status FlashService::tryAddLock(
